Refactor RealmMyTeam companion methods to TeamsRepository
Extracted RealmMyTeam companion methods into TeamsRepository interface and TeamsRepositoryImpl.
Added new methods: getTeamsByTeamId, getTeamsByUserId, getTeamCreator, getResourceIdsByUser, insertTeams, insertTeam.
Moved serialization logic to TeamsRepository companion object.
Updated TransactionSyncManager and SyncManager to use TeamsRepository for team synchronization, removing reflection-based calls.
Refactored TeamCoursesFragment and TeamCoursesAdapter to fetch team creator asynchronously and remove direct Realm usage in Adapter.
Removed RealmMyTeam companion object and unused RealmMyLibrary methods.
Updated ServiceModule to provide TeamsRepository to sync managers.
